bookly_api/pyproject.toml
[tool.poetry]
name = "bookly_api"
version = "0.1.0"
description = "API for recommender system"
authors = ["dmitriyl99 <dmitriyl1899@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
alembic = "^1.7.4"
SQLAlchemy = "^1.4.25"
databases = "^0.5.3"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
